requirements/dev-to-freeze.txt
asgiref
brotlipy
certifi
cffi
chardet
coverage
dj-database-url
Django
django-appconf
django-compressor
django-markdownx>=4.0.0b1
django-sass-processor
idna
importlib-metadata
libsass
lxml
Markdown
psycopg2-binary
pycparser
pytz
PyYAML
rcssmin
requests
rjsmin
six
sqlparse
urllib3
whitenoise
zipp
